refactor(#1120): split tasks.rs (4,698 LOC) into tasks/ module directory
Decompose the monolithic tasks.rs into a well-structured module
directory following the same pattern as the inbox/ split (PR #1111):

- mod.rs (413 LOC): Task struct, TaskStore, core data layer
  (list_all, load_by_id, record_to_task, dependency eval, migration)
- handler.rs (682 LOC): MCP handle() dispatcher
- orphan.rs (513 LOC): orphan detection, health diagnostics,
  reconciliation
- sweep.rs (263 LOC): board-hygiene sweep internals (scan + emit)
- acl.rs (121 LOC): ACL predicates (can_mutate_task/record,
  system identity allow-list)
- tests.rs (2,703 LOC): all 75 existing tests, unchanged

All 75 tests pass, clippy clean, spawn_rationale_audit passes.


Agend-Agent: fixup-dev
Agend-Task: t-20260524045252223063-36
Agend-Branch: refactor/1120-tasks-module-split
Agend-Issued-At: 2026-05-24T04:53:06.262085+00:00
